fix: resolve 15 PHPStan level-10 errors from t145 controller split
- Move check_session_permission and check_session_owner_permission from
  PermissionTrait into SessionController (the only caller). PermissionTrait
  was accessing $this->database on 7 controllers that don't have the property.
- Remove dead $settings property and constructor param from RestController —
  all Settings usage in that class is via static methods.
- Remove RestController's unused $database property and constructor (no
  instance methods read it after the t145 split).
- Remove two stale AdminPage::SLUG suppressions from phpstan.neon — the class
  no longer exists and nothing references it.

PHPStan level 10: 15 errors → 0 errors.
